Remove artifact_component to reduce level of component nesting
