+++ wildlifecompliance/frontend/wildlifecompliance/src/components/internal/legal_case/legal_case.vue
@@ -563,7 +563,7 @@ export default {
             recordDescriptionHtml = this.cancelModalUrl(recordNumberElement)
         } else if (entity && entity.data_type === 'individual' && action === 'ok') {
             recordDescriptionHtml = this.insertPersonModalUrl({"entity": entity, "recordNumberElement": recordNumberElement})
-        } else if (entity && entity.data_type === 'artifact' && action === 'ok') {
+        } else if (entity && entity.data_type === 'document_artifact' && action === 'ok') {
             recordDescriptionHtml = this.insertArtifactModalUrl({"entity": entity, "recordNumberElement": recordNumberElement})
         } else if (entity && entity.data_type === 'url' && action === 'ok') {
             recordDescriptionHtml = this.insertModalUrl({"entity": entity, "recordNumberElement": recordNumberElement})
@@ -603,11 +603,12 @@ export default {
     */
     insertArtifactModalUrl: function({"entity": entity, "recordNumberElement": recordNumberElement}) {
         let replacementVal = ''
-            // TODO: replace with correct artifact url
-        if (entity.full_name) {
-            replacementVal = `<a contenteditable="false" target="_blank" href="/internal/users/${entity.id}">${entity.full_name}</a>`
+        // TODO: replace with correct artifact url
+        if (entity.artifact_type) {
+            replacementVal = `<a contenteditable="false" target="_blank" href="/internal/users/${entity.id}">${entity.artifact_type}</a>`
         }
         let recordDescriptionHtml = recordNumberElement[0].innerHTML.replace(this.tabSelectedKeyCombination, replacementVal).replace(/&nbsp\;/g, ' ');
+        console.log(recordDescriptionHtml);
         return recordDescriptionHtml;
     },
     cancelModalUrl: function(recordNumberElement) {
